feat: all pokemons page divded to smaller components
