+++ src/hooks/usePokemonsTable.ts
@@ -1,0 +1,48 @@
+import { useEffect, useState } from "react";
+import { fetchPokemons, type Pokemon } from "../api/fetchPokemons";
+import { sortPokemons } from "../utils/sortPokemons";
+
+export function usePokemonsTable() {
+  const [page, setPage] = useState(1);
+  const [pageSize, setPageSize] = useState(10);
+  const [pokemons, setPokemons] = useState<Pokemon[]>([]);
+  const [loading, setLoading] = useState(true);
+  const [searchValue, setSearchValue] = useState("");
+  const [filterValue, setFilterValue] = useState<string | null>(null);
+
+  useEffect(() => {
+    setLoading(true);
+    fetchPokemons(100).then((data) => {
+      setPokemons(data.results);
+      setLoading(false);
+    });
+  }, []);
+
+  let filteredPokemons = searchValue
+    ? pokemons.filter((pokemon) =>
+        pokemon.name.english.toLowerCase().includes(searchValue.toLowerCase())
+      )
+    : [...pokemons];
+
+  filteredPokemons = sortPokemons(filteredPokemons, filterValue);
+
+  const total = filteredPokemons.length;
+  const start = (page - 1) * pageSize;
+  const end = start + pageSize;
+  const pagePokemons = filteredPokemons.slice(start, end);
+
+  return {
+    page,
+    setPage,
+    pageSize,
+    setPageSize,
+    pokemons,
+    loading,
+    searchValue,
+    setSearchValue,
+    filterValue,
+    setFilterValue,
+    total,
+    pagePokemons,
+  };
+}
